修改 Bug, 回復新設計的分頁方式

+++ CTBC.CSFS.Models/WarningFraud.cs
@@ -11,7 +11,7 @@ namespace CTBC.CSFS.Models
         /// <summary>
         /// PK
         /// </summary>
-        public long No { get; set; }
+        public int No { get; set; }
         /// <summary>
         /// 
         /// </summary>
